+++ packages/dash-core-components/src/components/DatePickerSingle.react.js
@@ -55,6 +55,13 @@ DatePickerSingle.propTypes = {
     max_date_allowed: PropTypes.string,
 
     /**
+     * Specifies additional days between min_date_allowed and max_date_allowed
+     * that should be disabled. Accepted datetime.datetime objects or strings
+     * in the format 'YYYY-MM-DD'
+     */
+    disabled_days: PropTypes.arrayOf(PropTypes.string),
+
+    /**
      * Specifies the month that is initially presented when the user
      * opens the calendar. Accepts datetime.datetime objects or strings
      * in the format 'YYYY-MM-DD'
@@ -240,6 +247,7 @@ DatePickerSingle.defaultProps = {
     disabled: false,
     persisted_props: ['date'],
     persistence_type: 'local',
+    disabled_days: [],
 };
 
 export const propTypes = DatePickerSingle.propTypes;