+++ packages/dash-core-components/src/fragments/DatePickerRange.react.js
@@ -82,14 +82,20 @@ export default class DatePickerRange extends Component {
     }
 
     isOutsideRange(date) {
-        const {min_date_allowed, max_date_allowed} = convertToMoment(
-            this.props,
-            ['min_date_allowed', 'max_date_allowed']
-        );
+        const {
+            max_date_allowed,
+            min_date_allowed,
+            disabled_days,
+        } = convertToMoment(this.props, [
+            'max_date_allowed',
+            'min_date_allowed',
+            'disabled_days',
+        ]);
 
         return (
             (min_date_allowed && date.isBefore(min_date_allowed)) ||
-            (max_date_allowed && date.isAfter(max_date_allowed))
+            (max_date_allowed && date.isAfter(max_date_allowed)) ||
+            (disabled_days && disabled_days.some(d => date.isSame(d, 'day')))
         );
     }