+++ packages/dash-core-components/src/utils/convertToMoment.js
@@ -19,6 +19,8 @@ export default (newProps, momentProps) => {
                         undefined
                 );
             }
+        } else if (Array.isArray(value)) {
+            dest[key] = value.map(d => moment(d));
         } else {
             dest[key] = moment(value);